// web/src/api/models/SystemInterfaceStatus.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Routier API
 * Configuration and monitoring API for a Routier node.
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { BondstatStatus } from './BondstatStatus';
import {
    BondstatStatusFromJSON,
    BondstatStatusFromJSONTyped,
    BondstatStatusToJSON,
    BondstatStatusToJSONTyped,
} from './BondstatStatus';

/**
 * 
 * @export
 * @interface SystemInterfaceStatus
 */
export interface SystemInterfaceStatus {
    /**
     * 
     * @type {string}
     * @memberof SystemInterfaceStatus
     */
    addr?: string;
    /**
     * 
     * @type {BondstatStatus}
     * @memberof SystemInterfaceStatus
     */
    bond?: BondstatStatus;
    /**
     * 
     * @type {{ [key: string]: any; }}
     * @memberof SystemInterfaceStatus
     */
    link: { [key: string]: any; };
}

/**
 * Check if a given object implements the SystemInterfaceStatus interface.
 */
export function instanceOfSystemInterfaceStatus(value: object): value is SystemInterfaceStatus {
    if (!('link' in value) || value['link'] === undefined) return false;
    return true;
}

export function SystemInterfaceStatusFromJSON(json: any): SystemInterfaceStatus {
    return SystemInterfaceStatusFromJSONTyped(json, false);
}

export function SystemInterfaceStatusFromJSONTyped(json: any, ignoreDiscriminator: boolean): SystemInterfaceStatus {
    if (json == null) {
        return json;
    }
    return {
        
        'addr': json['addr'] == null ? undefined : json['addr'],
        'bond': json['bond'] == null ? undefined : BondstatStatusFromJSON(json['bond']),
        'link': json['link'],
    };
}

export function SystemInterfaceStatusToJSON(json: any): SystemInterfaceStatus {
    return SystemInterfaceStatusToJSONTyped(json, false);
}

export function SystemInterfaceStatusToJSONTyped(value?: SystemInterfaceStatus | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'addr': value['addr'],
        'bond': BondstatStatusToJSON(value['bond']),
        'link': value['link'],
    };
}